requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
Django==4.0.5
django-registration==3.3
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
